+++ packages/eslint-config/next.js
@@ -6,10 +6,12 @@ const project = resolve(process.cwd(), "tsconfig.json");
 module.exports = {
   extends: [
     "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
     "prettier",
     require.resolve("@vercel/style-guide/eslint/next"),
     "eslint-config-turbo",
   ],
+  parser: "@typescript-eslint/parser",
   globals: {
     React: true,
     JSX: true,